+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.6.0]
+
 ### Added
 
 - Add DNS resolver cache
@@ -74,7 +76,8 @@ All notable changes to this project will be documented in this file.
 
 - First public release
 
-[unreleased]: https://github.com/soosp/Host/compare/0.5.0...HEAD
+[unreleased]: https://github.com/soosp/Host/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/soosp/Host/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/soosp/Host/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/soosp/Host/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/soosp/Host/compare/0.2.1...0.3.0